chore(kms-connector): remove scheduler
