--- kms-connector/simple-connector/src/core/coordination/mod.rs
@@ -1,3 +1,0 @@
-pub mod scheduler;
-
-pub use scheduler::MessageScheduler;
